TASK313 - решение

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDAO.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface UserDAO {
     public List<User> getUsers();
-    public User getUser(int id);
+    public User getUser(Long id);
     public void createUser(User user);
-    public void updateUser(int id, User user);
-    public void deleteUser(int id);
+    public void updateUser(Long id, User user);
+    public void deleteUser(Long id);
     public User findByUsername(String email);
 }
